Refacotring create project
