+++ app/cli/createproject/createproject_cli_pvc_workflow.go
@@ -1,22 +1,16 @@
 package createproject
 
 import (
-	"errors"
 	"k8s-management-go/app/cli/dialogs"
 	"k8s-management-go/app/utils/logger"
 	"k8s-management-go/app/utils/loggingstate"
+	"k8s-management-go/app/utils/validator"
 )
 
-func ProjectWizardAskForExistingPersistentVolumeClaim() (pvcName string, err error) {
+func PersistentVolumeClaimWorkflow() (pvcName string, err error) {
 	log := logger.Log()
 	// Validator for pvc
-	validate := func(input string) error {
-		// a pvc name cannot be longer than 253 characters
-		if len(input) > 253 {
-			return errors.New("PVC name is too long! You can only use max. 253 characters. ")
-		}
-		return nil
-	}
+	validate := validator.ValidatePersistentVolumeClaim
 
 	// Prepare prompt
 	dialogs.ClearScreen()
@@ -24,7 +18,7 @@ func ProjectWizardAskForExistingPersistentVolumeClaim() (pvcName string, err err
 	// check if everything was ok
 	if err != nil {
 		loggingstate.AddErrorEntryAndDetails("  -> Unable to get persistent volume claim.", err.Error())
-		log.Errorf("[ProjectWizardAskForExistingPersistentVolumeClaim] Unable to get persistent volume claim. %s\n", err.Error())
+		log.Errorf("[PersistentVolumeClaimWorkflow] Unable to get persistent volume claim. %s\n", err.Error())
 		return pvcName, err
 	}
