+++ app/gui/uninstall/uninstall_gui_screen.go
@@ -46,7 +46,7 @@ func ScreenUninstall(window fyne.Window) fyne.CanvasObject {
 			} else {
 				models.AssignDryRun(false)
 			}
-			if !validator.ValidateNamespace(namespace) {
+			if !validator.ValidateNamespaceAvailableInConfig(namespace) {
 				namespaceErrorLabel.SetText("Error: namespace is unknown!")
 				namespaceErrorLabel.Show()
 				return
