+++ app/utils/validator/namespace_validator.go
@@ -1,13 +1,38 @@
 package validator
 
-import "k8s-management-go/app/models"
+import (
+	"errors"
+	"k8s-management-go/app/models"
+	"regexp"
+	"strings"
+)
 
 // check selected namespace against namespace list
-func ValidateNamespace(namespaceToValidate string) bool {
+func ValidateNamespaceAvailableInConfig(namespaceToValidate string) bool {
 	for _, ip := range models.GetIpConfiguration().Ips {
 		if ip.Namespace == namespaceToValidate {
 			return true
 		}
 	}
 	return false
+}
+
+// validate namespace
+func ValidateNewNamespace(input string) error {
+	// a namespace name cannot be longer than 63 characters
+	if len(input) > 63 {
+		return errors.New("Namespace name is too long! You can only use max. 63 characters. ")
+	}
+	// Regex to have DNS compatible string
+	regex := regexp.MustCompile(`^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$`)
+	if !regex.Match([]byte(input)) {
+		return errors.New("Namespace is not valid! It must fit to DNS specification! ")
+	}
+	// check, that namespace was not already used
+	for _, ipConfig := range models.GetIpConfiguration().Ips {
+		if strings.ToLower(ipConfig.Namespace) == strings.ToLower(input) {
+			return errors.New("Namespace already in use! ")
+		}
+	}
+	return nil
 }
